Add configured Vert.x Redis targets

Copilot-Session: e700cc87-e68c-470e-98c6-3776e2fecc94
