--- instrumentation/vertx/vertx-redis-client-4.0/javaagent/build.gradle.kts
@@ -1,39 +1,0 @@
-plugins {
-  id("otel.javaagent-instrumentation")
-}
-
-muzzle {
-  pass {
-    group.set("io.vertx")
-    module.set("vertx-redis-client")
-    versions.set("[4.0.0,)")
-    assertInverse.set(true)
-  }
-}
-
-dependencies {
-  library("io.vertx:vertx-redis-client:4.0.0")
-  compileOnly("io.vertx:vertx-codegen:4.0.0")
-
-  testInstrumentation(project(":instrumentation:netty:netty-4.1:javaagent"))
-
-  testLibrary("io.vertx:vertx-codegen:4.0.0")
-}
-
-tasks {
-  withType<Test>().configureEach {
-    usesService(gradle.sharedServices.registrations["testcontainersBuildService"].service)
-    systemProperty("collectMetadata", otelProps.collectMetadata)
-  }
-
-  val testStableSemconv = register<Test>("testStableSemconv") {
-    testClassesDirs = sourceSets.test.get().output.classesDirs
-    classpath = sourceSets.test.get().runtimeClasspath
-    jvmArgs("-Dotel.semconv-stability.opt-in=database,service.peer")
-    systemProperty("metadataConfig", "otel.semconv-stability.opt-in=database,service.peer")
-  }
-
-  check {
-    dependsOn(testStableSemconv)
-  }
-}
