--- instrumentation/vertx/vertx-redis-client-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/vertx/redisclient/v4_0/RedisStandaloneConnectionInstrumentation.java
@@ -1,232 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.vertx.redisclient.v4_0;
-
-import static io.opentelemetry.javaagent.instrumentation.vertx.redisclient.v4_0.VertxRedisClientSingletons.instrumenter;
-import static net.bytebuddy.matcher.ElementMatchers.isConstructor;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import io.vertx.core.Future;
-import io.vertx.core.net.NetSocket;
-import io.vertx.redis.client.RedisConnection;
-import io.vertx.redis.client.Request;
-import io.vertx.redis.client.Response;
-import io.vertx.redis.client.impl.RedisStandaloneConnection;
-import io.vertx.redis.client.impl.RedisURI;
-import io.vertx.redis.client.impl.RequestUtil;
-import java.util.List;
-import javax.annotation.Nullable;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.asm.Advice.AssignReturned;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-
-class RedisStandaloneConnectionInstrumentation implements TypeInstrumentation {
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("io.vertx.redis.client.impl.RedisStandaloneConnection");
-  }
-
-  @Override
-  public void transform(TypeTransformer transformer) {
-    transformer.applyAdviceToMethod(named("send"), getClass().getName() + "$SendAdvice");
-    transformer.applyAdviceToMethod(named("batch"), getClass().getName() + "$BatchAdvice");
-    transformer.applyAdviceToMethod(isConstructor(), getClass().getName() + "$ConstructorAdvice");
-  }
-
-  @SuppressWarnings("unused")
-  public static class SendAdvice {
-    public static class AdviceScope {
-      private final VertxRedisClientRequest otelRequest;
-      private final Context context;
-      private final Scope scope;
-
-      private AdviceScope(VertxRedisClientRequest otelRequest, Context context, Scope scope) {
-        this.otelRequest = otelRequest;
-        this.context = context;
-        this.scope = scope;
-      }
-
-      @Nullable
-      public static AdviceScope start(
-          RedisStandaloneConnection connection, @Nullable Request request, NetSocket netSocket) {
-
-        if (request == null) {
-          return null;
-        }
-
-        String commandName = VertxRedisClientSingletons.getCommandName(request.command());
-        RedisURI redisUri = VertxRedisClientSingletons.getRedisUri(connection);
-
-        VertxRedisClientRequest otelRequest =
-            new VertxRedisClientRequest(
-                commandName, RequestUtil.getArgs(request), redisUri, netSocket);
-        Context parentContext = Context.current();
-        if (!instrumenter().shouldStart(parentContext, otelRequest)) {
-          return null;
-        }
-        Context context = instrumenter().start(parentContext, otelRequest);
-        return new AdviceScope(otelRequest, context, context.makeCurrent());
-      }
-
-      @Nullable
-      public Future<Response> end(
-          @Nullable Future<Response> responseFuture, @Nullable Throwable throwable) {
-        scope.close();
-        if (throwable != null) {
-          instrumenter().end(context, otelRequest, null, throwable);
-        } else {
-          responseFuture =
-              VertxRedisClientSingletons.wrapEndSpan(responseFuture, context, otelRequest);
-        }
-        return responseFuture;
-      }
-    }
-
-    @Nullable
-    @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
-    public static AdviceScope onEnter(
-        @Advice.This RedisStandaloneConnection connection,
-        @Advice.Argument(0) @Nullable Request request,
-        @Advice.FieldValue("netSocket") NetSocket netSocket) {
-
-      return AdviceScope.start(connection, request, netSocket);
-    }
-
-    @Nullable
-    @AssignReturned.ToReturned
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class, inline = false)
-    public static Future<Response> onExit(
-        @Advice.Thrown @Nullable Throwable throwable,
-        @Advice.Return @Nullable Future<Response> responseFuture,
-        @Advice.Enter @Nullable AdviceScope adviceScope) {
-
-      if (adviceScope != null) {
-        return adviceScope.end(responseFuture, throwable);
-      }
-
-      return responseFuture;
-    }
-  }
-
-  @SuppressWarnings("unused")
-  public static class BatchAdvice {
-    public static class BatchAdviceScope {
-      private static final boolean VERTX_4_0_X = isVertx40x();
-
-      private final VertxRedisClientRequest otelRequest;
-      private final Context context;
-      private final Scope scope;
-      private final boolean emptyBatch;
-
-      private BatchAdviceScope(
-          VertxRedisClientRequest otelRequest, Context context, Scope scope, boolean emptyBatch) {
-        this.otelRequest = otelRequest;
-        this.context = context;
-        this.scope = scope;
-        this.emptyBatch = emptyBatch;
-      }
-
-      // Vert.x Redis 4.0.0 through 4.0.3 create a promise for an empty batch but never process a
-      // response that could complete it. Version 4.1.0.Beta1 added explicit empty-batch completion.
-      // The same 4.1.0.Beta1 update replaced the connection provider API, so init(RedisConnection)
-      // is a reliable marker for the affected 4.0.x line.
-      private static boolean isVertx40x() {
-        try {
-          return Class.forName(
-                      "io.vertx.redis.client.impl.RedisConnectionManager$RedisConnectionProvider",
-                      false,
-                      RedisStandaloneConnection.class.getClassLoader())
-                  .getDeclaredMethod("init", RedisConnection.class)
-              != null;
-        } catch (ReflectiveOperationException ignored) {
-          return false;
-        }
-      }
-
-      @Nullable
-      public static BatchAdviceScope start(
-          RedisStandaloneConnection connection,
-          @Nullable List<Request> requests,
-          NetSocket netSocket) {
-
-        if (requests == null) {
-          return null;
-        }
-
-        RedisURI redisUri = VertxRedisClientSingletons.getRedisUri(connection);
-
-        VertxRedisClientRequest otelRequest =
-            VertxRedisClientRequest.createBatch(requests, redisUri, netSocket);
-        Context parentContext = Context.current();
-        if (!instrumenter().shouldStart(parentContext, otelRequest)) {
-          return null;
-        }
-        Context context = instrumenter().start(parentContext, otelRequest);
-        return new BatchAdviceScope(
-            otelRequest, context, context.makeCurrent(), requests.isEmpty());
-      }
-
-      @Nullable
-      public Future<List<Response>> end(
-          @Nullable Future<List<Response>> responseFuture, @Nullable Throwable throwable) {
-        scope.close();
-        if (throwable != null) {
-          instrumenter().end(context, otelRequest, null, throwable);
-        } else if (emptyBatch
-            && VERTX_4_0_X
-            && responseFuture != null
-            && !responseFuture.isComplete()) {
-          // Vert.x 4.0.x never completes the promise returned for an empty batch.
-          instrumenter().end(context, otelRequest, null, null);
-        } else {
-          responseFuture =
-              VertxRedisClientSingletons.wrapEndSpan(responseFuture, context, otelRequest);
-        }
-        return responseFuture;
-      }
-    }
-
-    @Nullable
-    @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
-    public static BatchAdviceScope onEnter(
-        @Advice.This RedisStandaloneConnection connection,
-        @Advice.Argument(0) @Nullable List<Request> requests,
-        @Advice.FieldValue("netSocket") NetSocket netSocket) {
-
-      return BatchAdviceScope.start(connection, requests, netSocket);
-    }
-
-    @Nullable
-    @AssignReturned.ToReturned
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class, inline = false)
-    public static Future<List<Response>> onExit(
-        @Advice.Thrown @Nullable Throwable throwable,
-        @Advice.Return @Nullable Future<List<Response>> responseFuture,
-        @Advice.Enter @Nullable BatchAdviceScope adviceScope) {
-
-      if (adviceScope != null) {
-        return adviceScope.end(responseFuture, throwable);
-      }
-
-      return responseFuture;
-    }
-  }
-
-  @SuppressWarnings("unused")
-  public static class ConstructorAdvice {
-    @Advice.OnMethodExit(suppress = Throwable.class, inline = false)
-    public static void onExit(@Advice.This RedisStandaloneConnection connection) {
-      // used in 4.1.0, for 4.0.0 it is set in RedisConnectionProviderInstrumentation
-      VertxRedisClientSingletons.setRedisUri(
-          connection, VertxRedisClientSingletons.getRedisUriThreadLocal());
-    }
-  }
-}
