+++ instrumentation/vertx/vertx-redis-client/vertx-redis-client-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/vertx/redisclient/v4_0/VertxRedisClientSingletons.java
@@ -1,0 +1,123 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.vertx.redisclient.v4_0;
+
+import static io.opentelemetry.instrumentation.api.incubator.semconv.db.internal.DbExceptionEventExtractors.setDbClientExceptionEventExtractor;
+
+import io.opentelemetry.api.GlobalOpenTelemetry;
+import io.opentelemetry.context.Context;
+import io.opentelemetry.context.Scope;
+import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientAttributesExtractor;
+import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientMetrics;
+import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientSpanNameExtractor;
+import io.opentelemetry.instrumentation.api.incubator.semconv.service.peer.ServicePeerAttributesExtractor;
+import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
+import io.opentelemetry.instrumentation.api.instrumenter.InstrumenterBuilder;
+import io.opentelemetry.instrumentation.api.instrumenter.SpanKindExtractor;
+import io.opentelemetry.instrumentation.api.util.VirtualField;
+import io.vertx.core.Future;
+import io.vertx.redis.client.Command;
+import io.vertx.redis.client.impl.RedisStandaloneConnection;
+import io.vertx.redis.client.impl.RedisURI;
+import java.util.concurrent.CompletableFuture;
+import javax.annotation.Nullable;
+
+public class VertxRedisClientSingletons {
+  private static final String INSTRUMENTATION_NAME = "io.opentelemetry.vertx-redis-client-4.0";
+  private static final Instrumenter<VertxRedisClientRequest, Void> instrumenter;
+
+  private static final ThreadLocal<RedisURI> redisUriThreadLocal = new ThreadLocal<>();
+  private static final VirtualField<Command, String> COMMAND_NAME =
+      VirtualField.find(Command.class, String.class);
+  private static final VirtualField<RedisStandaloneConnection, RedisURI> REDIS_URI =
+      VirtualField.find(RedisStandaloneConnection.class, RedisURI.class);
+
+  static {
+    VertxRedisClientAttributesGetter getter = new VertxRedisClientAttributesGetter();
+    // Redis semantic conventions don't follow the regular pattern of adding db.namespace to the
+    // span name.
+    VertxRedisClientAttributesGetter spanNameAttributesGetter =
+        new VertxRedisClientAttributesGetter() {
+          @Override
+          @Nullable
+          public String getDbNamespace(VertxRedisClientRequest request) {
+            return null;
+          }
+        };
+
+    InstrumenterBuilder<VertxRedisClientRequest, Void> builder =
+        Instrumenter.<VertxRedisClientRequest, Void>builder(
+                GlobalOpenTelemetry.get(),
+                INSTRUMENTATION_NAME,
+                DbClientSpanNameExtractor.create(spanNameAttributesGetter))
+            .addAttributesExtractor(DbClientAttributesExtractor.create(getter))
+            .addAttributesExtractor(new VertxRedisClientAttributesExtractor())
+            .addAttributesExtractor(
+                ServicePeerAttributesExtractor.create(getter, GlobalOpenTelemetry.get()))
+            .addOperationMetrics(DbClientMetrics.get());
+    setDbClientExceptionEventExtractor(builder);
+
+    instrumenter = builder.buildInstrumenter(SpanKindExtractor.alwaysClient());
+  }
+
+  public static Instrumenter<VertxRedisClientRequest, Void> instrumenter() {
+    return instrumenter;
+  }
+
+  public static <T> Future<T> wrapEndSpan(
+      Future<T> future, Context context, VertxRedisClientRequest request) {
+    Context parentContext = Context.current();
+    CompletableFuture<T> result = new CompletableFuture<>();
+    future
+        .toCompletionStage()
+        .whenComplete(
+            (value, throwable) -> {
+              instrumenter().end(context, request, null, throwable);
+              try (Scope ignore = parentContext.makeCurrent()) {
+                if (throwable != null) {
+                  result.completeExceptionally(throwable);
+                } else {
+                  result.complete(value);
+                }
+              }
+            });
+    return Future.fromCompletionStage(result);
+  }
+
+  @Nullable
+  public static RedisURI getRedisUriThreadLocal() {
+    return redisUriThreadLocal.get();
+  }
+
+  public static void setRedisUriThreadLocal(RedisURI redisUri) {
+    redisUriThreadLocal.set(redisUri);
+  }
+
+  public static void clearRedisUriThreadLocal() {
+    redisUriThreadLocal.remove();
+  }
+
+  public static void setCommandName(Command command, String commandName) {
+    COMMAND_NAME.set(command, commandName);
+  }
+
+  @Nullable
+  public static String getCommandName(Command command) {
+    return COMMAND_NAME.get(command);
+  }
+
+  public static void setRedisUri(
+      RedisStandaloneConnection connection, @Nullable RedisURI redisUri) {
+    REDIS_URI.set(connection, redisUri);
+  }
+
+  @Nullable
+  public static RedisURI getRedisUri(RedisStandaloneConnection connection) {
+    return REDIS_URI.get(connection);
+  }
+
+  private VertxRedisClientSingletons() {}
+}
