+++ instrumentation/vertx/vertx-redis-client/vertx-redis-client-4.4.5/metadata.yaml
@@ -1,0 +1,6 @@
+display_name: Vert.x Redis Client
+description: >
+  This instrumentation records the target a Vert.x Redis client was configured with for the versions
+  that build their connections from Redis connect options. Spans and metrics themselves come from the
+  Vert.x Redis Client 4.0 instrumentation.
+library_link: https://vertx.io/docs/vertx-redis-client/java/
