+++ settings.gradle.kts
@@ -764,7 +764,10 @@ include(":instrumentation:vertx:vertx-kafka-client-3.6:testing")
 include(":instrumentation:vertx:vertx-kafka-client-3.6:vertx-kafka-client-3.6-testing")
 include(":instrumentation:vertx:vertx-kafka-client-3.6:vertx-kafka-client-4-testing")
 include(":instrumentation:vertx:vertx-kafka-client-3.6:vertx-kafka-client-5-testing")
-include(":instrumentation:vertx:vertx-redis-client-4.0:javaagent")
+include(":instrumentation:vertx:vertx-redis-client:vertx-redis-client-4.0:javaagent")
+include(":instrumentation:vertx:vertx-redis-client:vertx-redis-client-4.0:javaagent-unit-tests")
+include(":instrumentation:vertx:vertx-redis-client:vertx-redis-client-4.4.5:javaagent")
+include(":instrumentation:vertx:vertx-redis-client:vertx-redis-client-4.4.5:javaagent-unit-tests")
 include(":instrumentation:vertx:vertx-rx-java-3.5:javaagent")
 include(":instrumentation:vertx:vertx-sql-client:vertx-sql-client-4.0:javaagent")
 include(":instrumentation:vertx:vertx-sql-client:vertx-sql-client-5.0:javaagent")
